# cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/certificates:
        get:
            tags:
                - LcmMtlsCertificateService
            description: mTLS Certificate Management (admin only)
            operationId: LcmMtlsCertificateService_ListMtlsCertificates
            parameters:
                - name: status
                  in: query
                  schema:
                    enum:
                        - MTLS_CERTIFICATE_STATUS_UNSPECIFIED
                        - MTLS_CERTIFICATE_STATUS_ACTIVE
                        - MTLS_CERTIFICATE_STATUS_EXPIRED
                        - MTLS_CERTIFICATE_STATUS_REVOKED
                        - MTLS_CERTIFICATE_STATUS_SUSPENDED
                    type: string
                    format: enum
                - name: clientId
                  in: query
                  schema:
                    type: string
                - name: issuerName
                  in: query
                  schema:
                    type: string
                - name: commonName
                  in: query
                  schema:
                    type: string
                - name: includeExpired
                  in: query
                  schema:
                    type: boolean
                - name: includeRevoked
                  in: query
                  schema:
                    type: boolean
                - name: tenantId
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMtlsCertificatesResponse'
        post:
            tags:
                - LcmMtlsCertificateService
            operationId: LcmMtlsCertificateService_IssueMtlsCertificate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IssueMtlsCertificateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IssueMtlsCertificateResponse'
    /v1/certificates/{serialNumber}:
        get:
            tags:
                - LcmMtlsCertificateService
            operationId: LcmMtlsCertificateService_GetMtlsCertificate
            parameters:
                - name: serialNumber
                  in: path
                  required: true
                  schema:
                    type: string
                - name: fingerprintSha256
                  in: query
                  schema:
                    type: string
                - name: commonName
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMtlsCertificateResponse'
        put:
            tags:
                - LcmMtlsCertificateService
            operationId: LcmMtlsCertificateService_UpdateMtlsCertificate
            parameters:
                - name: serialNumber
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMtlsCertificateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateMtlsCertificateResponse'
        delete:
            tags:
                - LcmMtlsCertificateService
            operationId: LcmMtlsCertificateService_DeleteMtlsCertificate
            parameters:
                - name: serialNumber
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteMtlsCertificateResponse'
...
    /v1/mtls-certificate-requests:
        get:
            tags:
                - LcmMtlsCertificateRequestService
            description: mTLS Certificate Request Management (admin only)
            operationId: LcmMtlsCertificateRequestService_ListMtlsCertificateRequests
            parameters:
                - name: status
                  in: query
                  schema:
                    enum:
                        - MTLS_CERTIFICATE_REQUEST_STATUS_UNSPECIFIED
                        - MTLS_CERTIFICATE_REQUEST_STATUS_PENDING
                        - MTLS_CERTIFICATE_REQUEST_STATUS_APPROVED
                        - MTLS_CERTIFICATE_REQUEST_STATUS_REJECTED
                        - MTLS_CERTIFICATE_REQUEST_STATUS_ISSUED
                        - MTLS_CERTIFICATE_REQUEST_STATUS_CANCELLED
                    type: string
                    format: enum
                - name: clientId
                  in: query
                  schema:
                    type: string
                - name: issuerName
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMtlsCertificateRequestsResponse'
        post:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_CreateMtlsCertificateRequest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMtlsCertificateRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateMtlsCertificateRequestResponse'
    /v1/mtls-certificate-requests/{data.id}:
        put:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_UpdateMtlsCertificateRequest
            parameters:
                - name: data.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMtlsCertificateRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateMtlsCertificateRequestResponse'
    /v1/mtls-certificate-requests/{id}:
        get:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_GetMtlsCertificateRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: requestId
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMtlsCertificateRequestResponse'
        delete:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_DeleteMtlsCertificateRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteMtlsCertificateRequestResponse'
    /v1/mtls-certificate-requests/{id}/approve:
        post:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_ApproveMtlsCertificateRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApproveMtlsCertificateRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApproveMtlsCertificateRequestResponse'
    /v1/mtls-certificate-requests/{id}/reject:
        post:
            tags:
                - LcmMtlsCertificateRequestService
            operationId: LcmMtlsCertificateRequestService_RejectMtlsCertificateRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RejectMtlsCertificateRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RejectMtlsCertificateRequestResponse'
...
components:
    schemas:
...
        ApproveMtlsCertificateRequestRequest:
            required:
                - id
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                issuerName:
                    type: string
                validityDays:
                    type: integer
                    format: int32
                notes:
                    type: string
            description: Approve mTLS Certificate Request
        ApproveMtlsCertificateRequestResponse:
            type: object
            properties:
                request:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
                certificate:
                    $ref: '#/components/schemas/ClientCertificate'
...
        CreateMtlsCertificateRequestRequest:
            required:
                - clientId
                - commonName
            type: object
            properties:
                clientId:
                    type: string
                commonName:
                    type: string
                csrPem:
                    type: string
                publicKey:
                    type: string
                dnsNames:
                    type: array
                    items:
                        type: string
                ipAddresses:
                    type: array
                    items:
                        type: string
                issuerName:
                    type: string
                certType:
                    enum:
                        - MTLS_CERT_TYPE_UNSPECIFIED
                        - MTLS_CERT_TYPE_CLIENT
                        - MTLS_CERT_TYPE_INTERNAL
                    type: string
                    format: enum
                validityDays:
                    type: integer
                    format: int32
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                notes:
                    type: string
            description: Create mTLS Certificate Request
        CreateMtlsCertificateRequestResponse:
            type: object
            properties:
                mtlsCertificateRequest:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
...
        DeleteMtlsCertificateRequestResponse:
            type: object
            properties:
                mtlsCertificateRequest:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
        DeleteMtlsCertificateResponse:
            type: object
            properties:
                mtlsCertificate:
                    $ref: '#/components/schemas/MtlsCertificate'
...
        GetMtlsCertificateRequestResponse:
            type: object
            properties:
                mtlsCertificateRequest:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
        GetMtlsCertificateResponse:
            type: object
            properties:
                mtlsCertificate:
                    $ref: '#/components/schemas/MtlsCertificate'
...
        IssueMtlsCertificateRequest:
            required:
                - clientId
                - commonName
            type: object
            properties:
                clientId:
                    type: string
                commonName:
                    type: string
                csrPem:
                    type: string
                publicKeyPem:
                    type: string
                dnsNames:
                    type: array
                    items:
                        type: string
                ipAddresses:
                    type: array
                    items:
                        type: string
                issuerName:
                    type: string
                certType:
                    enum:
                        - MTLS_CERT_TYPE_UNSPECIFIED
                        - MTLS_CERT_TYPE_CLIENT
                        - MTLS_CERT_TYPE_INTERNAL
                    type: string
                    format: enum
                validityDays:
                    type: integer
                    format: int32
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                notes:
                    type: string
            description: Issue mTLS Certificate (create new certificate directly, bypassing request flow)
        IssueMtlsCertificateResponse:
            type: object
            properties:
                mtlsCertificate:
                    $ref: '#/components/schemas/MtlsCertificate'
                caCertificatePem:
                    type: string
...
        ListMtlsCertificateRequestsResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MtlsCertificateRequest'
                total:
                    type: string
        ListMtlsCertificatesResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/MtlsCertificate'
                total:
                    type: string
...
        RejectMtlsCertificateRequestRequest:
            required:
                - id
                - reason
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                reason:
                    type: string
            description: Reject mTLS Certificate Request
        RejectMtlsCertificateRequestResponse:
            type: object
            properties:
                mtlsCertificateRequest:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
...
        UpdateMtlsCertificateRequest:
            required:
                - serialNumber
            type: object
            properties:
                serialNumber:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                notes:
                    type: string
                updateMask:
                    type: string
                    format: field-mask
            description: Update mTLS Certificate (metadata, notes only - cert content is immutable)
        UpdateMtlsCertificateRequestRequest:
            required:
                - data
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
                updateMask:
                    type: string
                    format: field-mask
                allowMissing:
                    type: boolean
            description: Update mTLS Certificate Request
        UpdateMtlsCertificateRequestResponse:
            type: object
            properties:
                mtlsCertificateRequest:
                    $ref: '#/components/schemas/MtlsCertificateRequest'
        UpdateMtlsCertificateResponse:
            type: object
            properties:
                mtlsCertificate:
                    $ref: '#/components/schemas/MtlsCertificate'
...