# gen/go/lcm/service/v1/mtls_certificate_request_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: lcm/service/v1/mtls_certificate_request.proto

package servicev1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLcmMtlsCertificateRequestServiceApproveMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/ApproveMtlsCertificateRequest"
const OperationLcmMtlsCertificateRequestServiceCreateMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/CreateMtlsCertificateRequest"
const OperationLcmMtlsCertificateRequestServiceDeleteMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/DeleteMtlsCertificateRequest"
const OperationLcmMtlsCertificateRequestServiceGetMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/GetMtlsCertificateRequest"
const OperationLcmMtlsCertificateRequestServiceListMtlsCertificateRequests = "/lcm.service.v1.LcmMtlsCertificateRequestService/ListMtlsCertificateRequests"
const OperationLcmMtlsCertificateRequestServiceRejectMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/RejectMtlsCertificateRequest"
const OperationLcmMtlsCertificateRequestServiceUpdateMtlsCertificateRequest = "/lcm.service.v1.LcmMtlsCertificateRequestService/UpdateMtlsCertificateRequest"

type LcmMtlsCertificateRequestServiceHTTPServer interface {
	ApproveMtlsCertificateRequest(context.Context, *ApproveMtlsCertificateRequestRequest) (*ApproveMtlsCertificateRequestResponse, error)
	CreateMtlsCertificateRequest(context.Context, *CreateMtlsCertificateRequestRequest) (*CreateMtlsCertificateRequestResponse, error)
	DeleteMtlsCertificateRequest(context.Context, *DeleteMtlsCertificateRequestRequest) (*DeleteMtlsCertificateRequestResponse, error)
	GetMtlsCertificateRequest(context.Context, *GetMtlsCertificateRequestRequest) (*GetMtlsCertificateRequestResponse, error)
	// ListMtlsCertificateRequests mTLS Certificate Request Management (admin only)
	ListMtlsCertificateRequests(context.Context, *ListMtlsCertificateRequestsRequest) (*ListMtlsCertificateRequestsResponse, error)
	RejectMtlsCertificateRequest(context.Context, *RejectMtlsCertificateRequestRequest) (*RejectMtlsCertificateRequestResponse, error)
	UpdateMtlsCertificateRequest(context.Context, *UpdateMtlsCertificateRequestRequest) (*UpdateMtlsCertificateRequestResponse, error)
}

func RegisterLcmMtlsCertificateRequestServiceHTTPServer(s *http.Server, srv LcmMtlsCertificateRequestServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/mtls-certificate-requests", _LcmMtlsCertificateRequestService_ListMtlsCertificateRequests0_HTTP_Handler(srv))
	r.GET("/v1/mtls-certificate-requests/{id}", _LcmMtlsCertificateRequestService_GetMtlsCertificateRequest0_HTTP_Handler(srv))
	r.POST("/v1/mtls-certificate-requests", _LcmMtlsCertificateRequestService_CreateMtlsCertificateRequest0_HTTP_Handler(srv))
	r.PUT("/v1/mtls-certificate-requests/{data.id}", _LcmMtlsCertificateRequestService_UpdateMtlsCertificateRequest0_HTTP_Handler(srv))
	r.DELETE("/v1/mtls-certificate-requests/{id}", _LcmMtlsCertificateRequestService_DeleteMtlsCertificateRequest0_HTTP_Handler(srv))
	r.POST("/v1/mtls-certificate-requests/{id}/approve", _LcmMtlsCertificateRequestService_ApproveMtlsCertificateRequest0_HTTP_Handler(srv))
	r.POST("/v1/mtls-certificate-requests/{id}/reject", _LcmMtlsCertificateRequestService_RejectMtlsCertificateRequest0_HTTP_Handler(srv))
}

func _LcmMtlsCertificateRequestService_ListMtlsCertificateRequests0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMtlsCertificateRequestsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceListMtlsCertificateRequests)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMtlsCertificateRequests(ctx, req.(*ListMtlsCertificateRequestsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMtlsCertificateRequestsResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_GetMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMtlsCertificateRequestRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceGetMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMtlsCertificateRequest(ctx, req.(*GetMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_CreateMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateMtlsCertificateRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceCreateMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateMtlsCertificateRequest(ctx, req.(*CreateMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_UpdateMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMtlsCertificateRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceUpdateMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMtlsCertificateRequest(ctx, req.(*UpdateMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_DeleteMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteMtlsCertificateRequestRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceDeleteMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteMtlsCertificateRequest(ctx, req.(*DeleteMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_ApproveMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ApproveMtlsCertificateRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceApproveMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ApproveMtlsCertificateRequest(ctx, req.(*ApproveMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ApproveMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

func _LcmMtlsCertificateRequestService_RejectMtlsCertificateRequest0_HTTP_Handler(srv LcmMtlsCertificateRequestServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RejectMtlsCertificateRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLcmMtlsCertificateRequestServiceRejectMtlsCertificateRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RejectMtlsCertificateRequest(ctx, req.(*RejectMtlsCertificateRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RejectMtlsCertificateRequestResponse)
		return ctx.Result(200, reply)
	}
}

type LcmMtlsCertificateRequestServiceHTTPClient interface {
	ApproveMtlsCertificateRequest(ctx context.Context, req *ApproveMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *ApproveMtlsCertificateRequestResponse, err error)
	CreateMtlsCertificateRequest(ctx context.Context, req *CreateMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *CreateMtlsCertificateRequestResponse, err error)
	DeleteMtlsCertificateRequest(ctx context.Context, req *DeleteMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *DeleteMtlsCertificateRequestResponse, err error)
	GetMtlsCertificateRequest(ctx context.Context, req *GetMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *GetMtlsCertificateRequestResponse, err error)
	// ListMtlsCertificateRequests mTLS Certificate Request Management (admin only)
	ListMtlsCertificateRequests(ctx context.Context, req *ListMtlsCertificateRequestsRequest, opts ...http.CallOption) (rsp *ListMtlsCertificateRequestsResponse, err error)
	RejectMtlsCertificateRequest(ctx context.Context, req *RejectMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *RejectMtlsCertificateRequestResponse, err error)
	UpdateMtlsCertificateRequest(ctx context.Context, req *UpdateMtlsCertificateRequestRequest, opts ...http.CallOption) (rsp *UpdateMtlsCertificateRequestResponse, err error)
}

type LcmMtlsCertificateRequestServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLcmMtlsCertificateRequestServiceHTTPClient(client *http.Client) LcmMtlsCertificateRequestServiceHTTPClient {
	return &LcmMtlsCertificateRequestServiceHTTPClientImpl{client}
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) ApproveMtlsCertificateRequest(ctx context.Context, in *ApproveMtlsCertificateRequestRequest, opts ...http.CallOption) (*ApproveMtlsCertificateRequestResponse, error) {
	var out ApproveMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests/{id}/approve"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceApproveMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) CreateMtlsCertificateRequest(ctx context.Context, in *CreateMtlsCertificateRequestRequest, opts ...http.CallOption) (*CreateMtlsCertificateRequestResponse, error) {
	var out CreateMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceCreateMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) DeleteMtlsCertificateRequest(ctx context.Context, in *DeleteMtlsCertificateRequestRequest, opts ...http.CallOption) (*DeleteMtlsCertificateRequestResponse, error) {
	var out DeleteMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceDeleteMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) GetMtlsCertificateRequest(ctx context.Context, in *GetMtlsCertificateRequestRequest, opts ...http.CallOption) (*GetMtlsCertificateRequestResponse, error) {
	var out GetMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceGetMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListMtlsCertificateRequests mTLS Certificate Request Management (admin only)
func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) ListMtlsCertificateRequests(ctx context.Context, in *ListMtlsCertificateRequestsRequest, opts ...http.CallOption) (*ListMtlsCertificateRequestsResponse, error) {
	var out ListMtlsCertificateRequestsResponse
	pattern := "/v1/mtls-certificate-requests"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceListMtlsCertificateRequests))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) RejectMtlsCertificateRequest(ctx context.Context, in *RejectMtlsCertificateRequestRequest, opts ...http.CallOption) (*RejectMtlsCertificateRequestResponse, error) {
	var out RejectMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests/{id}/reject"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceRejectMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *LcmMtlsCertificateRequestServiceHTTPClientImpl) UpdateMtlsCertificateRequest(ctx context.Context, in *UpdateMtlsCertificateRequestRequest, opts ...http.CallOption) (*UpdateMtlsCertificateRequestResponse, error) {
	var out UpdateMtlsCertificateRequestResponse
	pattern := "/v1/mtls-certificate-requests/{data.id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLcmMtlsCertificateRequestServiceUpdateMtlsCertificateRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}